requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard
blessed==1.19.0
Brotli==1.1.0
celery
certifi==2021.10.8
cffi==1.16.0
charset-normalizer==2.0.7
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
cron-descriptor==1.4.0
cryptography==41.0.5
cssselect2==0.7.0
Django==4.2.4
django-celery
django-celery-beat
django-countries==7.5.1
django-debug-toolbar==4.2.0
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-mathfilters==1.0.0
django-mptt==0.15.0
django-mptt-admin==2.5.1
django-rosetta
django-timezone-field==6.0.1
fire==0.4.0
flake8==6.1.0
flower==2.0.1
fonttools==4.43.1
halo==0.0.31
html5lib==1.1
humanize==4.8.0
idna==3.2
kombu==5.3.1
log-symbols==0.0.14
mccabe==0.7.0
pathspec==0.9.0
paypal-checkout-serversdk==1.0.3
paypalhttp==1.0.1
Pillow==10.1.0
polib==1.2.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pydyf==0.8.0
pyflakes==3.1.0
Pygments==2.10.0
pyOpenSSL==23.3.0
pyphen==0.14.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-editor==1.0.4
python-slugify==8.0.1
pytz==2021.3
readchar==2.0.1
regex==2023.10.3
requests
requests-toolbelt==0.9.1
runflare==1.1.6
six==1.16.0
spinners==0.0.24
sqlparse==0.4.4
termcolor==1.1.0
text-unidecode==1.3
tinycss2==1.2.1
tornado==6.3.3
typing_extensions==4.7.1
tzdata
tzlocal==3.0
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
weasyprint==60.1
webencodings==0.5.1
websockets==10.0
zopfli==0.2.3
